Sample Python Apps/requirements.txt
flask
opencensus-ext-azure